+++ gateway/llm_client/__init__.py
@@ -1,5 +1,5 @@
 from .schemas import REQUIRED_OUTPUT_KEYS, validate_output_schema
-from .prompts import SYSTEM_PROMPT, build_user_prompt
+from .prompts import SYSTEM_PROMPT, build_repair_prompt, build_user_prompt
 from .mock_llm import MockLLM
 from .validator import validate_response, ValidationResult
 from .client import LLMClient, get_client
@@ -8,6 +8,7 @@ __all__ = [
     "REQUIRED_OUTPUT_KEYS",
     "validate_output_schema",
     "SYSTEM_PROMPT",
+    "build_repair_prompt",
     "build_user_prompt",
     "MockLLM",
     "validate_response",